+++ tests/compile/num_compose.rs
@@ -1,0 +1,11 @@
+use compose_idents::compose_idents;
+
+compose_idents!(my_fn = [foo, _, 1, _, bar]; {
+    fn my_fn() -> u32 {
+        42
+    }
+});
+
+fn main() {
+    assert_eq!(foo_1_bar(), 42);
+}